Created contrast profile csv files
